requirements.txt
asgiref==3.8.1
Django==5.1.6
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
et_xmlfile==2.0.0
Markdown==3.7
openpyxl==3.1.5
pillow==11.1.0
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.1
